requirements.txt
pymongo==3.11.2
tqdm==4.51.0
paramiko==2.7.2
xmltodict==0.12.0
pandas==1.1.3
requests==2.24.0
python_dateutil==2.8.1
